Add support for Azure
Previously, the code was designed solely for local disk execution. This
commit introduces Azure support, enabling the code to run on the cloud.
To make this possible, include your account name and token in the
Const.h file before compiling the code. To run the code on the local
disk, include the newly introduced -local option.